DOCS/STRIPE/webhook/requirements.txt
Flask>=2.0
stripe>=6.0
python-dotenv>=1.0
